Refs #31. Moved and fixed some for ui.
